lambda/requirements.txt
pytesseract==0.3.10
Pillow==10.2.0
openai==1.12.0  # for GPT-4 API
pytest==7.4.3  # for testing
black==23.12.1  # for formatting 
